добавлен smart компонент
