+++ src/view/smart.js
@@ -1,0 +1,25 @@
+import Abstract from '../view/abstract';
+
+export default class Smart extends Abstract {
+  constructor() {
+    super();
+    this._data = {};
+  }
+
+  _updateData(update) {
+    this._data = Object.assign({}, this._data, update);
+  }
+
+  _updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+    const newElement = this.getElement();
+    parent.replaceChild(newElement, prevElement);
+    this._restoreHandlers();
+  }
+
+  _restoreHandlers() {
+    throw new Error(`Abstract method not implemented`);
+  }
+}
